fix(diarize): sf.read s dtype=float32 — 3× manje memorije po epizodi
sf.read bez dtype vraća float64, pa .float() radi drugu kopiju: 3× memorije
(3h WAV ~2 GB umjesto 0.7 GB; 20h ~14 GB umjesto 4.6 GB). S float32 je
from_numpy bez kopije, a .float() no-op.

Ne diram prosljeđivanje waveforma umjesto putanje — komentar u diarize.py kaže
da je to svjestan workaround za AudioDecoder. Putanja JEST provjereno ispravna na
pyannote 4.0.4 (sabor_pipeline/02_diarize.py), ali to je drugi kod-put i traži
zasebnu validaciju prije nego uđe u nightly.

Motiv je nightly: manji otisak znači manje swapanja na sistemskom disku, koji je
na ovom stroju usko grlo (19 GB slobodno, swap 5/6 GB zauzet).

+++ colab_diarize/diarize_canary.py
@@ -204,7 +204,9 @@ def run_diarization(pipeline, wav_file, min_speakers=None, max_speakers=None):
     import soundfile as sf
 
     # Učitaj audio
-    data, sample_rate = sf.read(wav_file)
+    # dtype="float32" — bez njega sf.read vraća float64 pa .float() radi drugu kopiju
+    # (3× memorije). Vidi isti komentar u diarize.py; mjereno 2026-08-25.
+    data, sample_rate = sf.read(wav_file, dtype="float32")
     waveform = torch.from_numpy(data).float().unsqueeze(0)
 
     # Parametri
